+++ lib/screens/experience.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:personal_website/components/stepper_item.dart';
+import 'package:personal_website/components/stepper.dart';
 import 'package:personal_website/controller/experience.dart';
 import 'package:personal_website/models/experience.dart';
 import 'package:personal_website/screens/screen_wrapper.dart';
@@ -36,30 +36,43 @@ class ExperienceScreen extends GetResponsiveView<ResponsiveController> {
           const SizedBox(
             height: 30,
           ),
-          for (int experienceIdx = 0;
-              experienceIdx < experiences.length;
-              experienceIdx++) ...[
-            StepperItem(
-              isFirst: experienceIdx == 0,
-              height: elementHeight,
-              activeBarColor: Colors.blue,
-              barWidth: 2,
-              experience: experiences[experienceIdx],
-              wide: wideMode,
-            ),
-            for (int jobIdx = 0;
-                jobIdx < experiences[experienceIdx].jobs.length;
-                jobIdx++)
-              SubStepperItem(
-                isLast: ((experienceIdx == experiences.length - 1) &
-                    (jobIdx == experiences[experienceIdx].jobs.length - 1)),
-                activeBarColor: Colors.blue,
-                barWidth: 2,
-                gap: elementHeight,
-                job: experiences[experienceIdx].jobs[jobIdx],
-                wide: wideMode,
-              ),
-          ]
+          ListView.builder(
+              shrinkWrap: true,
+              itemCount: experiences.length,
+              itemBuilder: (BuildContext context, int experienceIdx) {
+                return ExperienceStepper(
+                  wideMode: wideMode,
+                  elementHeight: elementHeight,
+                  experience: experiences[experienceIdx],
+                  experiencesLength: experiences.length,
+                  index: experienceIdx,
+                );
+              }),
+
+          // for (int experienceIdx = 0;
+          //     experienceIdx < experiences.length;
+          //     experienceIdx++) ...[
+          //   StepperItem(
+          //     isFirst: experienceIdx == 0,
+          //     height: elementHeight,
+          //     activeBarColor: Colors.blue,
+          //     barWidth: 2,
+          //     experience: experiences[experienceIdx],
+          //     wide: wideMode,
+          //   ),
+          //   for (int jobIdx = 0;
+          //       jobIdx < experiences[experienceIdx].jobs.length;
+          //       jobIdx++)
+          //     SubStepperItem(
+          //       isLast: ((experienceIdx == experiences.length - 1) &
+          //           (jobIdx == experiences[experienceIdx].jobs.length - 1)),
+          //       activeBarColor: Colors.blue,
+          //       barWidth: 2,
+          //       gap: elementHeight,
+          //       job: experiences[experienceIdx].jobs[jobIdx],
+          //       wide: wideMode,
+          //     ),
+          // ]
         ]);
       }),
     ));